Update FileInfo when the file is committed
Fixes https://github.com/Alluxio/alluxio/issues/13128
When this guard is added the issue does not happen anymore.

The trade-off is, when we `getFileInfo` on a file that is currently
being written, the number of `BlockId` and `FileBlockInfo` we get will
mismatch. I don't think that is a problem.

pr-link: Alluxio/alluxio#13131
change-id: cid-12fd9fa47ca7149ff9b6a1b5c29fe43907945342