Refactor BlockMaster interface to avoid copying the lost block set
This is a better approach of achieving
https://github.com/Alluxio/alluxio/pull/13454

The current implementation copies the `DefaultBlockMaster.mLostBlocks`
to whoever is using the set. This change improves the `BlockMaster`
interface to avoid the need to copy that collection. The `mLostBlocks`
set can be huge when there are workers lost. Also the cost is very high
because we just copy on every read.

pr-link: Alluxio/alluxio#13467
change-id: cid-7d1a2a9eaa1f0e957c1455bbbc9b7910202776d6